Rename solution and namespaces
