+++ HRSS/HRSS.csproj
@@ -8,8 +8,8 @@
     <ProjectGuid>{F6FB7CAD-7FBA-448B-98D7-B20FA24600BF}</ProjectGuid>
     <OutputType>WinExe</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>TestScreenshot</RootNamespace>
-    <AssemblyName>TestScreenshot</AssemblyName>
+    <RootNamespace>HRSS</RootNamespace>
+    <AssemblyName>HRSS</AssemblyName>
     <TargetFrameworkVersion>v4.0</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <FileUpgradeFlags>
