+++ HRSS/NativeMethods.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Runtime.InteropServices;
 
-namespace TestScreenshot
+namespace HRSS
 {
     [System.Security.SuppressUnmanagedCodeSecurity()]
     internal sealed class NativeMethods
