+++ HRSS/Program.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Windows.Forms;
 
-namespace TestScreenshot
+namespace HRSS
 {
     static class Program
     {
