+++ HRSS/Properties/Settings.settings
@@ -1,5 +1,5 @@
 ﻿<?xml version='1.0' encoding='utf-8'?>
-<SettingsFile xmlns="http://schemas.microsoft.com/VisualStudio/2004/01/settings" CurrentProfile="(Default)" GeneratedClassNamespace="TestScreenshot.Properties" GeneratedClassName="Settings">
+<SettingsFile xmlns="http://schemas.microsoft.com/VisualStudio/2004/01/settings" CurrentProfile="(Default)" GeneratedClassNamespace="HRSS.Properties" GeneratedClassName="Settings">
   <Profiles />
   <Settings>
     <Setting Name="Setting" Type="System.String" Scope="Application">
