+++ HRSS/Properties/app.config
@@ -2,15 +2,15 @@
 <configuration>
     <configSections>
         <sectionGroup name="applicationSettings" type="System.Configuration.ApplicationSettingsGroup, System, Version=2.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" >
-            <section name="TestScreenshot.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=2.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" requirePermission="false" />
+            <section name="HRSS.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=2.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" requirePermission="false" />
         </sectionGroup>
     </configSections>
     <applicationSettings>
-        <TestScreenshot.Properties.Settings>
+        <HRSS.Properties.Settings>
             <setting name="Setting" serializeAs="String">
                 <value>test</value>
             </setting>
-        </TestScreenshot.Properties.Settings>
+        </HRSS.Properties.Settings>
     </applicationSettings>
     <system.windows.forms jitDebugging="true" />
 </configuration>
